pq metrics: IORatioMetric edge-case logging, use int at interface
By using `int` type in `IORatioMetric#incrementBy(int,int)`, we simplify the
failure scenarios while still allowing the desired behaviour, since this is
always called in practice with `byte[]#length`.

We ensure that attempts to decrement the value are ignored, and result in a
log message, and that overflows reduce precision and are also logged.

Together, these ensure that long overflows won't ever result in pipeline
crashes.

+++ logstash-core/src/main/java/org/logstash/ackedqueue/IORatioMetric.java
@@ -11,7 +11,7 @@ interface IORatioMetric extends UserMetric<Double>, org.logstash.instrument.metr
 
     Value getLifetime();
 
-    void incrementBy(long bytesIn, long bytesOut);
+    void incrementBy(int bytesIn, int bytesOut);
 
     @Override
     default MetricType getType() {
@@ -37,7 +37,7 @@ interface IORatioMetric extends UserMetric<Double>, org.logstash.instrument.metr
         }
 
         @Override
-        public void incrementBy(long bytesIn, long bytesOut) {
+        public void incrementBy(int bytesIn, int bytesOut) {
             // no-op
         }
